pyproject.toml
[tool.poetry]
name = "CHHiP-UQ"
version = "0.1.0"
description = "Python packages for the Uncertainty Quantification Project"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
loguru = "^0.7.2"
black = "^24.4.2"
matplotlib = "^3.9.1"
cytoolz = "^0.12.3"
ipykernel = "^6.29.5"
fastexcel = "^0.11.5"
jupyter-black = "^0.3.4"
isort = "^5.13.2"
pathos = "^0.3.2"
notebook = "^7.2.1"
jupyter = "^1.0.0"
rt-utils = "^1.2.7"
tqdm = "^4.66.4"
simpleitk = "^2.3.1"
nibabel = "^5.2.1"
pydicom = "^2.4.4"
h5py = "^3.11.0"
pytest-mock = "^3.14.0"
pydot = "^3.0.1"
graphviz = "^0.20.3"
torch = "^2.4.0"
torchvision = "^0.19.0"
torchviz = "^0.0.2"
torchio = { version = "^0.18.90", python = ">=3.12,<3.13" }
lightning = "^2.4.0"
tensorboardx = "^2.6.2.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
